+++ weblate/trans/tests/test_component.py
@@ -112,6 +112,12 @@ class ComponentTest(RepoTestCase):
         component = self.create_ts_mono()
         self.verify_component(component, 2, 'cs', 4)
 
+    def test_create_appstore(self):
+        component = self.create_appstore()
+        self.verify_component(
+            component, 1, 'en-US', 3, 'Weblate - continuous localization'
+        )
+
     def test_create_po_pot(self):
         component = self._create_component(
             'po',